% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{argsetter}
\alias{argsetter}
\title{Setear defaults de argumentos para una función, a nivel de sesión}
\usage{
argsetter(fn, defaults)
}
\arguments{
\item{fn}{Una función.}

\item{defaults}{Lista (o vector nombrado) con los nuevos valores por default.}
}
\value{
La misma función, con los formals modificados.
}
\description{
Setear defaults de argumentos para una función, a nivel de sesión
}
\examples{
f <- function(a = 1, b = 2) a + b
f <- argsetter(f, list(a = 10))
f()  # 12
}
