% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cdp}
\alias{cdp}
\title{Navigate up n levels in the directory tree}
\usage{
cdp(n = 1)
}
\arguments{
\item{n}{Integer. Number of levels to go up. Default \code{1}.}
}
\value{
Invisibly returns the new working directory path.
}
\description{
Repeatedly applies \code{dirname()} to the current working directory
and then changes into the resulting path. Equivalent to running \code{cd ../..}
\code{n} times in a shell.
}
\examples{
\dontrun{
cdp()    # go up one level
cdp(3)   # go up three levels
}

}
