% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dv}
\alias{dv}
\title{Alias corto para \code{list()}}
\usage{
dv()
}
\arguments{
\item{...}{Elementos a pasar a \code{list()}.}
}
\value{
Un objeto \code{list}.
}
\description{
Wrapper con nombre corto para \code{\link{list}}, pensado para uso
interactivo en la consola donde escribir \code{list(...)} repetidamente
es tedioso.
}
\examples{
dv(a = 1, b = 2)
}
